Extract brush world validation
